Fix absolute paths on Windows with file scheme
This commit starts adding support for building and running WAGI on
Windows. Specifically, it:

- updates the wasi-experimental-http-wasmtime crate to v0.4, which now
compiles and runs on Windows
- handles absolute paths with file scheme when defining a module

Please note that at this time, absolute paths without the `file:` scheme
will cause an issue, something to keep in mind as we decide removing
the legacy support for scheme-less modules.

Signed-off-by: Radu M <root@radu.sh>