feat(*): Adds per module log files
This splits out logging stderr from the main server process into their own files.
It does nothing to expose those files somewhere else, but the paths to the logs
can be reconstructed using the given config and the new `id` method on the
`Method` struct. We could also look at doing some sort of in memory buffer or
stream in the future and swap this out for the file. This PR also makes a change to
ensure that the list of modules in a `ModuleConfig` is now unique using a `HashSet`.

I also did some campground rule cleanup here with a whole slew of clippy warnings
and removal of unnecessary clones.

Closes #44
Closes #7