chore: update to `wasmtime` 0.35.3
This is the most recent version of `wasmtime` that supports module linking.
Later versions have removed module linking in favor of the component model, but
the component model work is still in progress as of this writing.

This required updating `wasi-experimental-http-wasmtime` as well, and this
required minor changes to accomodate the new API.

There are also some minor whitespace changes due to `rustfmt` having its way.

Signed-off-by: Joel Dice <joel.dice@fermyon.com>